# py_sample/generate-schema-use-openapi-generator/app/openapi.yaml
openapi: 3.0.3
info:
  license:
    name: MIT
  title: Swagger Petstore
  version: 1.0.0
servers:
- description: local
  url: http://localhost:8000
paths:
  /pets:
    get:
      operationId: listPets
      parameters:
      - description: How many items to return at one time (max 100)
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pets'
          description: A paged array of pets
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
      summary: List all pets
      tags:
      - pets
    post:
      operationId: createPets
      responses:
        "201":
          description: Null response
      summary: Create a pet
      tags:
      - pets
  /pets/{petId}:
    get:
      operationId: showPetById
      parameters:
      - description: The id of the pet to retrieve
        explode: false
        in: path
        name: petId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pets'
          description: Expected response to a valid request
      summary: Info for a specific pet
      tags:
      - pets
components:
  schemas:
    Pet:
      example:
        name: name
        id: 0
        tag: tag
      properties:
        id:
          format: int64
          title: id
          type: integer
        name:
          title: name
          type: string
        tag:
          title: tag
          type: string
      required:
      - id
      - name
      title: Pet
    Pets:
      items:
        $ref: '#/components/schemas/Pet'
      type: array
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message