# py_sample/generate-schema-use-openapi-generator/app/src/openapi_server/models/pet.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class Pet(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Pet - a model defined in OpenAPI

        id: The id of this Pet.
        name: The name of this Pet.
        tag: The tag of this Pet [Optional].
    """

    id: int = Field(alias="id")
    name: str = Field(alias="name")
    tag: Optional[str] = Field(alias="tag", default=None)

Pet.update_forward_refs()